requirements.txt
Django>=1.11,<1.12
djangorestframework>=3.6,<3.7
markdown # For browseble API
psycopg2

# Django extensions
whitenoise
django-admin-interface
drfdocs